worker/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "grpcio-tools"]
build-backend = "setuptools.build_meta"

[project]
name = "llmserve_worker"
version = "0.1.0"
dependencies = [
    "grpcio>=1.71.0",
    "grpcio-tools>=1.71.0",
    "typer[all]>=0.15.4",
    "loguru>=0.7.3",
    "torch>=2.4.0",
    "numpy>=2.2.0",
    "transformers>4.47.1",
    "flashinfer-python>=0.2.5",
    "typer>=0.15.4",
]

[project.scripts]
llmserve-worker = "llmserve_worker.cli:app"

[tool.setuptools]
packages = ["llmserve_worker"]
